+++ apps/platform/public/svg/shared/index.ts
@@ -17,6 +17,7 @@ import RolesSVG from './roles.svg'
 import CheckmarkSVG from './checkmark.svg'
 import TrashWhiteSVG from './trash-white.svg'
 import RollbackSVG from './rollback.svg'
+import MinusSquareSVG from './minus-square.svg'
 import EyeOpenSVG from './eye.svg'
 import EyeSlashSVG from './eye-slash.svg'
 import InfoSVG from './info.svg'
@@ -50,4 +51,5 @@ export {
   LockSVG,
   RegenerateSVG,
   TickCircleSVG
+  MinusSquareSVG,
 }