+++ apps/platform/src/app/(main)/(settings)/settings/@profile/page.tsx
@@ -10,6 +10,7 @@ import GeneralSettings from '@/components/userSettings/profile/generalSettings'
 import EmailSettings from '@/components/userSettings/profile/emailSettings'
 import DeleteProfile from '@/components/userSettings/profile/deleteProfile'
 import ApiKeySection from '@/components/userSettings/apiKeys/apiKeySection'
+import AuthProvider from '@/components/userSettings/profile/authProvider'
 
 function ProfilePage(): React.JSX.Element {
   const setUser = useSetAtom(userAtom)
@@ -31,8 +32,15 @@ function ProfilePage(): React.JSX.Element {
 
   return (
     <main className="flex flex-col gap-y-10">
+      <div className="pt-4">
+        <h1 className="text-2xl font-bold">Profile</h1>
+        <p className="font-medium text-white/60 mt-2.5">
+          Changes made to your profile will be applied to all of your workspaces.
+        </p>
+      </div>
       <GeneralSettings />
       <EmailSettings />
+      <AuthProvider />
       <Separator className="w-full bg-white/15" />
       <ApiKeySection />
       <Separator className="w-full bg-white/15" />