+++ apps/platform/src/store/index.ts
@@ -10,11 +10,15 @@ import type {
   User,
   Variable,
   VariableVersion,
+  WorkspaceRole,
   WorkspaceWithTierLimitAndProjectCount
 } from '@keyshade/schema'
 
 export const userAtom = atomWithStorage<Partial<User> | null>('user', null)
 
+export const allWorkspacesAtom = atom<WorkspaceWithTierLimitAndProjectCount[]>(
+  []
+)
 export const selectedWorkspaceAtom =
   atom<WorkspaceWithTierLimitAndProjectCount | null>(null)
 
@@ -66,6 +70,8 @@ export const environmentsOfProjectAtom = atom<
   GetAllEnvironmentsOfProjectResponse['items']
 >([])
 
+export const rolesOfWorkspaceAtom = atom<WorkspaceRole[]>([])
+
 export const selectedApiKeyAtom = atom<ApiKey | null>(null)
 export const apiKeysOfProjectAtom = atom<ApiKey[]>([])
 export const selectedProjectPrivateKeyAtom = atom<string | null>(null)
@@ -92,10 +98,17 @@ export const rollbackSecretOpenAtom = atom<boolean>(false)
 export const createEnvironmentOpenAtom = atom<boolean>(false)
 export const editEnvironmentOpenAtom = atom<boolean>(false)
 export const deleteEnvironmentOpenAtom = atom<boolean>(false)
+export const deleteWorkspaceOpenAtom = atom<boolean>(false)
 
 export const createApiKeyOpenAtom = atom<boolean>(false)
 export const editApiKeyOpenAtom = atom<boolean>(false)
 export const deleteApiKeyOpenAtom = atom<boolean>(false)
+export const apiKeyOneTimeDisplayDialogOpenAtom = atom<boolean>(false)
+export const oneTimeSecretValueAtom = atom<string>("")
+
+export const createRolesOpenAtom = atom<boolean>(false)
+export const editRolesOpenAtom = atom<boolean>(false)
+export const deleteRolesOpenAtom = atom<boolean>(false)
 
 export const deleteAccountOpenAtom = atom<boolean>(false)